+++ src/chart_types/heatmap/specs/heatmap.ts
@@ -20,6 +20,7 @@
 import React from 'react';
 
 import { ChartTypes } from '../..';
+import { Predicate } from '../../../common/predicate';
 import { ScaleType } from '../../../scales/constants';
 import { SeriesScales, Spec } from '../../../specs';
 import { SpecTypes } from '../../../specs/constants';
@@ -28,7 +29,6 @@ import { Accessor, AccessorFn } from '../../../utils/accessor';
 import { Color, Datum, RecursivePartial } from '../../../utils/common';
 import { config } from '../layout/config/config';
 import { Config } from '../layout/types/config_types';
-import { Predicate } from '../utils/common';
 
 const defaultProps = {
   chartType: ChartTypes.Heatmap,
@@ -69,7 +69,7 @@ export interface HeatmapSpec extends Spec {
   ySortPredicate: Predicate;
   xScaleType: SeriesScales['xScaleType'];
   config: RecursivePartial<Config>;
-  highlightedData?: { x: any[]; y: any[] };
+  highlightedData?: { x: Array<string | number>; y: Array<string | number> };
   name?: string;
 }