+++ src/chart_types/partition_chart/layout/viewmodel/fill_text_layout.ts
@@ -17,15 +17,6 @@
  * under the License.
  */
 
-import chroma from 'chroma-js';
-
-import {
-  colorIsDark,
-  combineColors,
-  getTextColorIfTextInvertible,
-  isColorValid,
-  makeHighContrastColor,
-} from '../../../../common/color_calcs';
 import { TAU } from '../../../../common/constants';
 import {
   Coordinate,
@@ -40,9 +31,8 @@ import {
 } from '../../../../common/geometry';
 import { logarithm } from '../../../../common/math';
 import { integerSnap, monotonicHillClimb } from '../../../../common/optimize';
-import { Box, Font, PartialFont, TextContrast, TextMeasure, VerticalAlignments } from '../../../../common/text_utils';
-import { Color, ValueFormatter } from '../../../../utils/common';
-import { Logger } from '../../../../utils/logger';
+import { Box, Font, PartialFont, TextMeasure, VerticalAlignments } from '../../../../common/text_utils';
+import { ValueFormatter } from '../../../../utils/common';
 import { Layer } from '../../specs';
 import { Config, Padding } from '../types/config_types';
 import {
@@ -113,7 +103,7 @@ function getVerticalAlignment(
     case VerticalAlignments.bottom:
       return -(container.y1 - linePitch * (totalRowCount - 1 - rowIndex) - paddingBottom - fontSize * overhang);
     default:
-      return -((container.y0 + container.y1) / 2 + (linePitch * (rowIndex - totalRowCount)) / 2);
+      return -((container.y0 + container.y1) / 2 + (linePitch * (rowIndex + 1 - totalRowCount)) / 2);
   }
 }
 
@@ -149,6 +139,7 @@ export const getRectangleRowGeometry: GetShapeRowGeometry<RectangleConstruction>
       maximumRowLength: 0,
     };
   }
+
   const rowAnchorY = getVerticalAlignment(
     container,
     verticalAlignment,
@@ -198,70 +189,18 @@ function getAllBoxes(
 ): Box[] {
   return rawTextGetter(node)
     .split(' ')
+    .filter(Boolean)
     .map((text) => ({ text, ...sizeInvariantFontShorthand }))
     .concat(
       valueFormatter(valueGetter(node))
         .split(' ')
+        .filter(Boolean)
         .map((text) => ({ text, ...sizeInvariantFontShorthand, ...valueFont })),
     );
 }
 
 function getWordSpacing(fontSize: number) {
   return fontSize / 4;
-}
-
-/**
- * Determine the color for the text hinging on the parameters of textInvertible and textContrast
- * @internal
- */
-export function getFillTextColor(
-  textColor: Color,
-  textInvertible: boolean,
-  textContrast: TextContrast,
-  sliceFillColor: string,
-  containerBackgroundColor?: Color,
-): string | undefined {
-  const bgColorAlpha = isColorValid(containerBackgroundColor) ? chroma(containerBackgroundColor).alpha() : 1;
-  if (!isColorValid(containerBackgroundColor) || bgColorAlpha < 1) {
-    if (bgColorAlpha < 1) {
-      Logger.expected('Text contrast requires a background color with an alpha value of 1', 1, bgColorAlpha);
-    } else if (containerBackgroundColor !== 'transparent') {
-      Logger.warn(`Invalid background color "${String(containerBackgroundColor)}"`);
-    }
-
-    return getTextColorIfTextInvertible(
-      colorIsDark(sliceFillColor),
-      colorIsDark(textColor),
-      textColor,
-      false,
-      'white', // never used
-    );
-  }
-
-  let adjustedTextColor: string | undefined = textColor;
-  const containerBackground = combineColors(sliceFillColor, containerBackgroundColor);
-  const textShouldBeInvertedAndTextContrastIsFalse = textInvertible && !textContrast;
-  const textShouldBeInvertedAndTextContrastIsSetToTrue = textInvertible && typeof textContrast !== 'number';
-  const textContrastIsSetToANumberValue = typeof textContrast === 'number';
-  const textShouldNotBeInvertedButTextContrastIsDefined = textContrast && !textInvertible;
-
-  // change the contrast for the inverted slices
-  if (textShouldBeInvertedAndTextContrastIsFalse || textShouldBeInvertedAndTextContrastIsSetToTrue) {
-    const backgroundIsDark = colorIsDark(combineColors(sliceFillColor, containerBackgroundColor));
-    const specifiedTextColorIsDark = colorIsDark(textColor);
-    adjustedTextColor = getTextColorIfTextInvertible(
-      backgroundIsDark,
-      specifiedTextColorIsDark,
-      textColor,
-      textContrast,
-      containerBackground,
-    );
-    // if textContrast is a number then take that into account or if textInvertible is set to false
-  } else if (textContrastIsSetToANumberValue || textShouldNotBeInvertedButTextContrastIsDefined) {
-    return makeHighContrastColor(adjustedTextColor, containerBackground);
-  }
-
-  return adjustedTextColor;
 }
 
 type GetShapeRowGeometry<C> = (
@@ -285,7 +224,6 @@ function fill<C>(
   shapeConstructor: ShapeConstructor<C>,
   getShapeRowGeometry: GetShapeRowGeometry<C>,
   getRotation: GetRotation,
-  containerBackgroundColor?: Color,
 ) {
   return function fillClosure(
     config: Config,
@@ -309,30 +247,12 @@ function fill<C>(
         ? VerticalAlignments.bottom
         : VerticalAlignments.top;
       const fontSizes = allFontSizes[Math.min(node.depth, allFontSizes.length) - 1];
-      const {
-        textColor,
-        textInvertible,
-        fontStyle,
-        fontVariant,
-        fontFamily,
-        fontWeight,
-        valueFormatter,
-        padding,
-        textContrast,
-        textOpacity,
-      } = {
+      const { fontStyle, fontVariant, fontFamily, fontWeight, valueFormatter, padding, textOpacity, clipText } = {
         ...fillLabel,
         valueFormatter: formatter,
         ...layer.fillLabel,
         ...layer.shape,
       };
-      const fillTextColor = getFillTextColor(
-        textColor,
-        textInvertible,
-        textContrast,
-        node.fillColor,
-        containerBackgroundColor,
-      );
 
       const valueFont = {
         ...fillLabel,
@@ -346,7 +266,7 @@ function fill<C>(
         fontVariant,
         fontWeight,
         fontFamily,
-        textColor,
+        textColor: node.textColor,
         textOpacity,
       };
       const allBoxes = getAllBoxes(rawTextGetter, valueGetter, valueFormatter, sizeInvariantFont, valueFont, node);
@@ -367,8 +287,9 @@ function fill<C>(
           cy,
           padding,
           node,
+          clipText,
         ),
-        fillTextColor,
+        fillTextColor: node.textColor,
       };
     };
   };
@@ -387,6 +308,7 @@ function tryFontSize<C>(
   node: ShapeTreeNode,
   boxes: Box[],
   maxRowCount: number,
+  clipText?: boolean,
 ) {
   return function tryFontSizeFn(initialRowSet: RowSet, fontSize: Pixels): { rowSet: RowSet; completed: boolean } {
     let rowSet: RowSet = initialRowSet;
@@ -421,6 +343,7 @@ function tryFontSize<C>(
         rotation,
         verticalAlignment,
         leftAlign,
+        clipText,
         rows: [...new Array(targetRowCount)].map(() => ({
           rowWords: [],
           rowAnchorX: NaN,
@@ -468,7 +391,7 @@ function tryFontSize<C>(
           const wordBeginning = currentRowLength;
           currentRowLength += currentBox.width + wordSpacing;
 
-          if (currentRowLength <= currentRow.maximumLength) {
+          if (clipText || currentRowLength <= currentRow.maximumLength) {
             currentRowWords.push({ ...currentBox, wordBeginning });
             currentRow.length = currentRowLength;
             measuredBoxes.shift();
@@ -501,6 +424,7 @@ function getRowSet<C>(
   cy: Coordinate,
   padding: Padding,
   node: ShapeTreeNode,
+  clipText: boolean,
 ) {
   const tryFunction = tryFontSize(
     measure,
@@ -515,6 +439,7 @@ function getRowSet<C>(
     node,
     boxes,
     maxRowCount,
+    clipText,
   );
 
   // find largest fitting font size
@@ -548,9 +473,8 @@ export function fillTextLayout<C>(
   shapeConstructor: ShapeConstructor<C>,
   getShapeRowGeometry: GetShapeRowGeometry<C>,
   getRotation: GetRotation,
-  containerBackgroundColor?: Color,
 ) {
-  const specificFiller = fill(shapeConstructor, getShapeRowGeometry, getRotation, containerBackgroundColor);
+  const specificFiller = fill(shapeConstructor, getShapeRowGeometry, getRotation);
   return function fillTextLayoutClosure(
     measure: TextMeasure,
     rawTextGetter: RawTextGetter,
@@ -607,7 +531,7 @@ export function fillTextLayout<C>(
           return {
             rowSets: [...rowSets, nextRowSet],
             fontSizes: fontSizes.map((layerFontSizes: Pixels[], index: number) =>
-              index === layerIndex && !layers[layerIndex]?.fillLabel?.maximizeFontSize
+              !isNaN(nextRowSet.fontSize) && index === layerIndex && !layers[layerIndex]?.fillLabel?.maximizeFontSize
                 ? layerFontSizes.filter((size: Pixels) => size <= nextRowSet.fontSize)
                 : layerFontSizes,
             ),