+++ src/chart_types/xy_chart/tooltip/tooltip.ts
@@ -68,7 +68,7 @@ export function getHighligthedValues(
 
 /** @internal */
 export function formatTooltip(
-  { color, value: { x, y, mark, accessor }, seriesIdentifier }: IndexedGeometry,
+  { color, value: { x, y, mark, accessor, datum }, seriesIdentifier }: IndexedGeometry,
   spec: BasicSeriesSpec,
   isHeader: boolean,
   isHighlighted: boolean,
@@ -106,5 +106,6 @@ export function formatTooltip(
     color,
     isHighlighted: isHeader ? false : isHighlighted,
     isVisible,
+    datum,
   };
 }