+++ src/utils/legend.ts
@@ -17,12 +17,13 @@
  * under the License.
  */
 
-import { Position } from './common';
+import { LegendPositionConfig } from '../specs/settings';
+import { LayoutDirection } from './common';
 
 /** @internal */
-export const isHorizontalLegend = (legendPosition: Position) =>
-  legendPosition === Position.Bottom || legendPosition === Position.Top;
+export const isHorizontalLegend = (legendPosition: LegendPositionConfig) =>
+  legendPosition.direction === LayoutDirection.Horizontal;
 
 /** @internal */
-export const isHierarchicalLegend = (flatLegend: boolean | undefined, legendPosition: Position) =>
+export const isHierarchicalLegend = (flatLegend: boolean | undefined, legendPosition: LegendPositionConfig) =>
   !flatLegend && !isHorizontalLegend(legendPosition);