+++ stories/small_multiples/1_grid.tsx
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
+import { action } from '@storybook/addon-actions';
 import { boolean } from '@storybook/addon-knobs';
 import React, { useState } from 'react';
 
@@ -92,6 +92,7 @@ export const Example = () => {
               },
             },
           }}
+          onBrushEnd={action('brushEvent')}
         />
         <Axis id="time" title="horizontal" position={Position.Bottom} gridLine={{ visible: false }} />
         <Axis